Bug Fix: Avoid Null Pointer Exceptions.

+++ src/net/sourceforge/fenixedu/domain/User.java
@@ -66,7 +66,9 @@ public class User extends User_Base {
     public void delete() {
 	for (; !getIdentifications().isEmpty(); getIdentifications().get(0).delete())
 	    ;
-	getLoginRequest().delete();
+	if (hasLoginRequest()) {
+	    getLoginRequest().delete();
+	}
 	super.setPerson(null);
 	removeRootDomainObject();
 	super.deleteDomainObject();
